backend/requirements.txt
python-dotenv==1.0.1
django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-redis==5.4.0 
psycopg2-binary==2.9.9 
pillow==10.0.1
django-cloudinary-storage==0.3.0 
pyTelegramBotAPI==4.17.0
gunicorn==22.0.0
prettytable==3.10.0
drf-yasg==1.21.7 